4.15.2024 - grabbing stuff once more

+++ java/src/main/java/com/techelevator/dao/ReviewDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ReviewDao {
 
     List<Review> listByBeerId (int beerId);
-    Review getById(long reviewId);
+    Review getById(int reviewId);
     Review add(Review review, int beerId);
 //    Review update(Review review);
 //    void delete(int beerId);
