+++ java/src/main/java/com/techelevator/model/Review.java
@@ -2,13 +2,13 @@ package com.techelevator.model;
 
 public class Review {
 
-    private long reviewId;
+    private int reviewId;
     private int beerId;
     private String reviewer;
     private String review;
     private int rating;
 
-    public Review(long reviewId, int beerId, String reviewer, String review, int rating) {
+    public Review(int reviewId, int beerId, String reviewer, String review, int rating) {
         this.reviewId = reviewId;
         this.beerId = beerId;
         this.reviewer = reviewer;
@@ -24,7 +24,7 @@ public class Review {
         return reviewId;
     }
 
-    public void setReviewId(long reviewId) {
+    public void setReviewId(int reviewId) {
         this.reviewId = reviewId;
     }
